Avoid having Gists overwrite the document
Instead of inserting previews with .html() previews are now inserted
with .innerHTML.

When script tags are inserted with .html() jQuery will extract the tag
and insert it in the head of the page to execute it securely. This
avoids several things, including "Permission denied" in Internet
Explorer and multiple executions of the same script.

The downside is that Gists would overwrite the document when executed in
the head. With this change, script tags are now longer inserted in the
head, allowing Gists to behave nicely.