Introduced a customn jQuery plugin to manipulate text in textarea elements.
This addresses one of my major beefs with the built-in editor. I didn't like the way that the Bold button wouldn't toggle boldness on and off, and that blockquoting a set of lines would insert the > at the selection start point, and NOT the beginning of each selected line.

The jQuery TextManipulator plugin addresses this.

In your menubar button definitions, you can now do things like the following:

To add boldness, for instance, the search/replace keys are now replaced by the following:
    manipulate: ['selected', 'toggleWrap', '**']
This calls the toggleWrap function on the currently selected content, and wraps it with a **, making your markdown selection bold. Clicking it again will unwrap it.

You can do something like
    manipulate: ['selected', 'toggleWrap', '<div>', '</div>']
to set a custom suffix on your wrapper. If you leave the suffix blank, it will default to being the prefix, since that's more useful in this case.

To turn a bunch of lines into an unordered list, its search/replace is replaced by:
    manipulate: ['lines', 'toggleWrap', '* ']
This calls toggleWrap on the currently selected lines, and replaces the start of each line with '* ', making it a list.

You can do something like
    manipulate: ['selected', 'toggleWrap', '<span class="fancyline">', '</span>']
to set a custom suffix on your wrapper.
Unlike with selected content, if you leave the suffix blank, it will stay blank, since that's more useful in this case.

Things remaining to do:

 * Allow an arbitrary list of numbered items to be used as prefixes for Ordered Lists, instead of just "1. " like it is right now.
 * Improve the way that the Image and Link input boxes work. It'd be nice to be able to select from a list of wiki pages, right?
 * Try and work out how to get proper undo behaviour working for Firefox. Right now it will select the entire textarea content when you undo a menu bar action, and I don't think this is wanted.
 * See if there's something nice to be done with the code button - maybe detect if we're wrapping multiple lines and handle that differently?