Disable live preview on mobile/Android devices
The live editor behaves badly with mobile keyboards in that it inserts each
modification of the current word as a new word, e.g. when writing "Tea", what
will be written in the document is something like "TTeTea", getting even worse
when deleting characters afterwards, as a backspace will result in the total
result "TTeTeaTe".

Furthermore, if working in the more phone-like portrait mode, the horizontal
space is far too small for the two panes.