Update Mousetrap and use new 'mod' helper (#1245)
* Fix page list for files that have regexp special chars.

The page list collection logic was using the filename without any
scaping to create a regexp. This not only breaks for some names it might
even be a security problem by introducing bad regular expression as
filenames.

* Add another video to README.

* Test on Ruby 2.4

* Pass non-empty commit author details in transliteration test

Empty name or email are not allowed by libgit2 and cause a test failure
when the test suite is run against rugged_adapter.

* Solve bug when folder contain non-ascii character

When you create a file in a folder with non-ascii character, for exemple "Réseau", after creating the page, it throwed an "URI::InvalidURIError", given the fact that the url returned was "/Réseau/H%C3%A9y", only the part with the name of the file was correctly encoded.

So I propose to encode every part of the url to solve this issue
So I just

* readme: Use --document in place of deprecated options

* Release 4.1.0

* Skip tests for transliteration for adapters different than grit

* Fix date. Closes #1211

* Set bar_side for versioned pages. Closes #1226

* Update gemijione dependency. Closes #1227

* Release 4.1.1

* Update Mousetrap and use new 'mod' helper

Since version 1.4, a generic 'mod' helper can be used for cross
platform shortcuts. I updated Mousetrap to latest v1.6.1 and
used 'mod+s' instead of ['ctrl+s', 'command+s'] for the editor's
keyboard shortcut.

See https://craig.is/killing/mice#api.bind.combo