Tell cat_file the file mode (so it knows whether it's a symlink) instead of telling it the path for it to look up the mode itself.
In cases where we weren't looking up the mode from the path until now,
we're no worse off than we were before.  The only place we were
depending on this behavior was to work around a performance bottleneck
in `Porcelain.grep`.  However, in this case, the blob already knows
its mode and can provide it to `Porcelain.cat_file`.