Remove unused _layout.scss file
The SCSS being removed in this commit is not being imported or used.

To verify this, I added a `background-color: red;` definition to the
`#wiki-body` styles. On reload, the background colour did not change.
The background colour only changed after I added an `@import "_layout";`
call to the `template.scss.erb` file.

I also grepped through the compiled CSS files for any style definitions
that contained `#wiki-` and I did not find any.