Update Primer CSS to v15.2.0
This commit updates Primer to the latest version, v15.2.0. Primer is
only distributed as an NPM module. To update the library in place I had
to curl the compiled version from unpkg.com.

To reproduce my upgrade:

     curl https://unpkg.com/@primer/css@15.2.0/dist/primer.css > \
       lib/gollum/public/gollum/stylesheets/primer.css

I decided to include the library as a CSS file instead of a SCSS file.
Gollum's `primer.scss` file has historically not been SCSS anyway.

Because I am committing the compiled, vendor-provided CSS, we don't need
to compile it as SCSS. When I attempted to compile it as CSS, it was
very time consuming.