Use `md-` breakpoints to present wiki content
By wrapping the main wiki content and sidebar with a `container-lg` div,
and using `col-md-*` utilities to display both the wiki content and the
sidebar, we get a nicer, single-column display on mobile devices
(anything narrower than the `md-` breakpoint).

By using these Primer SCSS utilities, we also can do away with some
existing padding utilities, like the `pl-4` that was placed on the
`#wiki-body` div.