Float sidebar on the configured side
When we changed the way we render the sidebar with Primer utilities, we
also forced the sidebar to always be displayed to the right of the main
content.

We want to use the provided float utilities to float the main content
and sidebar in the right order. If you look at the Primer documentation,
[they actually suggest this method for "reversing" columns in a
grid][1].

If `bar_side` is right, then `body_side` is left, and vice versa.

[1]: https://primer.style/css/objects/grid#reversed-grid