Change wiki container padding on mobile
On mobile devices, there is no padding between the wiki content and the
edge of the screen. We can add horizontal padding with the `px-`
utility, and then set the padding back to `0` for screens larger than
the `md-` breakpoint using the `px-md-0` utility.

Similarly, there were no margins between the top of the mobile sidebar
and the top of the footer. I used the same technique to add margins
(`my-2`) for mobile and then set them back to `0` for larger screens.