If navbar is too wide, use overflow-x: scroll
If the navbar is too wide on the user's screen, this ensures that the
navbar can be scrolled horizontally.

It's unlikely that anyone will experience a too-wide navbar, especially
as I intend to add presentational navbar breakpoints in an upcoming
commit. However, this will be a better experience for those experiencing
this edge case.

This commit also picked up some whitespace issues. I am leaving them in
as I don't think it is too confusing.