Hide desktop nav items at `lg-`+ breakpoints
This commit uses `hide-*` utilities from Primer CSS to hide nav items at
`sm-` and `md-` breakpoints if they are items that will be displayed in
the mobile nav menu.

It also changes the presentation of buttons, by using the `buttom-sm`
class. It makes these buttons appear smaller across all breakpoints,
which I think is reasonable because the buttons are already quite large.

I also took this opportunity to deal with some whitespace issues in this
file, and reformat longer lines. This makes the HTML easier to scan and
will make future changes easier to diff.

Only the following nav items will appear on both desktop and mobile menus:

- Search bar
- "New Page" button