Use CSS classes instead of IDs for jQuery events
We now render the "New Page", "Rename Page", and "Upload Page" buttons
twice. Once in the mobile navigation menu and once in the desktop
navigation menu. This means that using CSS IDs is no longer valid. (IDs
should only ever be on one element on a page.)

If we want to use JavaScript events to trigger functionality when these
buttons are pressed, we need to use classes instead of IDs.