Change content x-axis padding for mobile devices
This tweaks the outer spacing of the wiki content container on small- to
medium-sized screens. We want some x-axis padding so there is space
between the text and the edge of the mobile device's screen.

On large 0 padding is fine, because the view width is larger than the
max wiki content container size.