Test page `page_header` options
In a previous commit, I made `#title` into a private method, because I
noticed it was not being used anywhere in templates.

This commit fixes a test that failed due to this method being changed to
a private method. What we really should be testing is the `#page_header`
method, which is used publicly.

It also adds additional tests that explicitly test all of the common
`#page_header` functionality that might be seen on the frontend.