Ensure "copy to clipboard" is usable
Modern browsers implement a permissions API that disallows websites from
writing to and reading the system clipboard without permission. To learn
more see the MDN's article about this[1].

This commit provides a basic check for clipboard access, so we don't
spam the browser or its console with errors accidentally.

We also don't want to run any of this code if no codeblocks are present.
So we can check for codeblocks, as well, and potentially exit early.

While I was refactoring, I also found it cumbersome to use a JS event
to find an element and its sibling element. So I've change the `event`
argument of the `copyToClipboard` function to be the codeblock ID we
generate, which lets us easily get the button for the codeblock as well
as the codeblock itself without a big event object.

[1]: https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/Interact_with_the_clipboard