Change Docker build test strategy
There was an issue where the Docker instance was not starting correctly.
This led to some hard-to-debug GitHub Actions CI logs

Before this commit, we depended on `docker run` to complete before
starting to run tests. This is why you see the `&` backgrounding command
and the `sleep 10` in the diff.

With `docker run --detach`, no further commands run until Docker exits,
after the process has successfully started (or failed).

So, we can just run tests after we're sure it's running.

Because a successfully running Docker instance is a requirement of the
Capybara tests working at all, I moved them into the same CI run
command. This way, if the first thing fails, the second thing also fails
and there are fewer GitHub Actions logs to sift through while
debugging.