Integration test_localization.rb: add wait times.
Integration tests were failing consistently on CI and hit-or-miss locally. This suggested a wait time issue. Indeed, it looks as if in some cases the search form was being submitted before the 'send_keys' operation was completed. Similarly, some assertions were being done before the page was reloaded.

Added explicit wait times to fix this.