Workflow precompile-assets: ignore .gz files when comparing changes
The logic for determining whether precompilation has changed any files, and thus whether a new commit is necessary, was flawed. .gz files are not identical for identical source files, so there was always a difference between the hashes calculated before and after asset compilation. This commit excludes .gz files from the hash calculation to remedy this.